Update ng-backstretch.js
Update directive to use two way bindings instead of expressions; make duration and fade optional; previously scope.duration would get initialized to 5000 no matter what was passed in for duration and scope.fade would get initialized to 1 no matter what was passed in for fade.  Directive now respects properties passed in but defaults duration and fade to 5000 and 1 respectively if values are not provided.